refactor(proof): 공통 dto 위치 변경
