+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/create/proof/CreateProofUseCase.java
@@ -1,5 +1,7 @@
 package me.jinjjahalgae.domain.proof.usecase.create.proof;
 
+import me.jinjjahalgae.domain.proof.usecase.create.common.ProofCreateRequest;
+
 public interface CreateProofUseCase {
 
     void execute(ProofCreateRequest request, Long contractId, Long userId);
