+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/create/reproof/CreateReProofUseCase.java
@@ -1,6 +1,6 @@
 package me.jinjjahalgae.domain.proof.usecase.create.reproof;
 
-import me.jinjjahalgae.domain.proof.usecase.create.proof.ProofCreateRequest;
+import me.jinjjahalgae.domain.proof.usecase.create.common.ProofCreateRequest;
 
 public interface CreateReProofUseCase {
