+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getlist/contractorlist/ContractorProofListResponse.java
@@ -1,6 +1,7 @@
 package me.jinjjahalgae.domain.proof.usecase.getlist.contractorlist;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import me.jinjjahalgae.domain.proof.usecase.getlist.common.ProofSimpleResponse;
 
 import java.time.LocalDateTime;
 /**
